Handle the strange TravisCI definition of failure
TravisCI provides no way to configure the test commands list to stop
on the first failure.  All listed commands will always be run to
completion - even if an earlier command exited with a failure code.

Their suggested way to prevent this is to build a shell pipeline
with "&&", instead I am just yanking all the logic back into the
Makefile - since make has a normal and reasonable interpretation
of "failure"

This allows earlier tests (like syntax tests) to give a quick and
immediate feedback if there is an issue - instead of waiting for
the hour-long full ubuntu install to complete.