Improve robustness of EFI boot variable handling
The system might have previously had an operating system installed on
it, so remove existing boot entries.  This is actually potentially
dangerous, but normal EFI bios boot configuration screens should allow
repair of any problems.  The problems are all restricted to booting
multiple operating systems, so this is the source of the comment that we
do not support that.

Additionally, allow each individual boot variable creation to fail,
since I have at least one motherboard with a broken boot variable.
This manifests as:
    Skipping unreadable variable "Boot0001": Input/output error
and causes the efibootmgr command to fail if it trys to use that boot
index to store a boot path.

Finally, to assist with debugging this in the future, print out a before
and after snapshot of the boot variables.