Stop using a mirror disk for the EFS
Limitations with most of the EFI booting tools means that using a mirror
as the storage for the EFS causes error messages.  This change switches
back to using a normal partition (and copies it to the redundant disks)

This also allows the use of the normal EFI boot variables management,
so my replacement code can be deleted.

The original purpose of the mirroring was to ensure that any current
or future updates to the contents of the EFS would be reflected in all
the redundant bootable partitions.